Move fields/inputs js & css is dedicate folder
