+++ tools/bazar/presentation/javascripts/fields/map-field-map-entry.js
