+++ tools/bazar/presentation/javascripts/inputs/checkbox-drag-and-drop.js
