+++ tools/bazar/presentation/javascripts/inputs/checkbox-tags.js
