+++ tools/bazar/presentation/javascripts/inputs/image-field.js
