+++ tools/bazar/presentation/styles/inputs/checkbox-drag-and-drop.css
