+++ tools/bazar/templates/inputs/checkbox_tags.twig
@@ -2,7 +2,7 @@
 
 {% block input %}
   {{ include_javascript('tools/tags/libs/vendor/bootstrap-tagsinput.min.js') }}
-  {{ include_javascript('tools/bazar/presentation/javascripts/bazar-tagsinput.js') }}
+  {{ include_javascript('tools/bazar/presentation/javascripts/inputs/checkbox-tags.js') }}
     <input 
       type="text" 
       name="{{ field.propertyName }}"
