+++ tools/bazar/templates/inputs/radio_tags.twig
@@ -1,8 +1,8 @@
 {% extends "@bazar/layouts/input.twig" %}
 
 {% block input %}
-  {{ include_javascript('tools/tags/libs/vendor/bootstrap-tagsinput.min.js') }}
-  {{ include_javascript('tools/bazar/presentation/javascripts/bazar-tagsinput.js') }}
+    {{ include_javascript('tools/tags/libs/vendor/bootstrap-tagsinput.min.js') }}
+    {{ include_javascript('tools/bazar/presentation/javascripts/inputs/checkbox-tags.js') }}
     {% if field.isDistantJson %}
       <div class="input-group">
     {% endif %}
