+++ tools/bazar/templates/inputs/user.twig
@@ -63,4 +63,4 @@
         </div>
     </div>
 {% endif %}
-{{ include_javascript('tools/bazar/presentation/javascripts/user-field-update-email.js') }}
+{{ include_javascript('tools/bazar/presentation/javascripts/inputs/user-field-update-email.js') }}
